[wip] training in 10human scenarios

+++ crowd_nav/train.py
@@ -33,15 +33,16 @@ def set_random_seeds(seed):
 def main(args):
     set_random_seeds(args.randomseed)
     # configure paths
-    make_new_dir = True
-    if os.path.exists(args.output_dir):
-        if args.overwrite:
-            shutil.rmtree(args.output_dir)
-        else:
-            shutil.rmtree(args.output_dir)
-    if make_new_dir:
-        os.makedirs(args.output_dir)
-        shutil.copy(args.config, os.path.join(args.output_dir, 'config.py'))
+    if args.resume is False:
+        make_new_dir = True
+        if os.path.exists(args.output_dir):
+            if args.overwrite:
+                shutil.rmtree(args.output_dir)
+            else:
+                shutil.rmtree(args.output_dir)
+        if make_new_dir:
+            os.makedirs(args.output_dir)
+            shutil.copy(args.config, os.path.join(args.output_dir, 'config.py'))
 
         # # insert the arguments from command line to the config file
         # with open(os.path.join(args.output_dir, 'config.py'), 'r') as fo:
@@ -170,13 +171,13 @@ def main(args):
     explorer = Explorer(env, robot, device, writer, memory, policy.gamma, target_policy=policy)
     policy.save_model(in_weight_file)
     # imitation learning
-    # if args.resume:
-    #     if not os.path.exists(rl_weight_file):
-    #         logging.error('RL weights does not exist')
-    #     policy.load_state_dict(torch.load(rl_weight_file))
-    #     model = policy.get_model()
-    #     rl_weight_file = os.path.join(args.output_dir, 'resumed_rl_model.pth')
-    #     logging.info('Load reinforcement learning trained weights. Resume training')
+    if args.resume:
+        if not os.path.exists(rl_weight_file):
+            logging.error('RL weights does not exist')
+        policy.load_state_dict(torch.load(rl_weight_file))
+        model = policy.get_model()
+        rl_weight_file = os.path.join(args.output_dir, 'resumed_rl_model.pth')
+        logging.info('Load reinforcement learning trained weights. Resume training')
     # elif os.path.exists(il_weight_file):
     #     policy.load_state_dict(torch.load(rl_weight_file))
     #     model = policy.get_model()
